Move network/constants to constants
