+++ src/constants/index.ts
@@ -4,3 +4,4 @@
 
 // Change this export to reconfigure constants
 export * from "./mainnet";
+export * from "./network";
