+++ src/constants/network.ts
